requirements.txt
tensorflow
streamlit
numpy
Pillow
gdown
